Refactor go to app action
